+++ src/Network/Network.Test/ScenarioTests/RouteTableTests.ps1
@@ -443,3 +443,62 @@ function Test-RouteTableWithDisableBgpRoutePropagation
         Clean-ResourceGroup $rgname
     }
 }
+
+<#
+.SYNOPSIS
+Tests RouteTableWithDisablePeeringRoute.
+#>
+function Test-RouteTableWithDisablePeeringRoute
+{
+    # Setup
+    $rgname = Get-ResourceGroupName
+    $routeTableName = Get-ResourceName
+    $rglocation = Get-ProviderLocation ResourceManagement
+    $resourceTypeParent = "Microsoft.Network/routeTables"
+    $location = Get-ProviderLocation $resourceTypeParent
+    
+    try 
+    {
+        # Create the resource group
+        $resourceGroup = New-AzResourceGroup -Name $rgname -Location $rglocation -Tags @{ testtag = "testval" } 
+        
+        # Create RouteTable with DisablePeeringRoute set to "All"
+        $rt = New-AzRouteTable -name $routeTableName -DisablePeeringRoute "All" -ResourceGroupName $rgname -Location $location
+
+        # Get RouteTable
+        $getRT = Get-AzRouteTable -name $routeTableName -ResourceGroupName $rgName
+        
+        #verification
+        Assert-AreEqual $rgName $getRT.ResourceGroupName
+        Assert-AreEqual $routeTableName $getRT.Name
+        Assert-AreEqual "All" $getRT.DisablePeeringRoute
+        Assert-NotNull $getRT.Etag
+        Assert-AreEqual 0 @($getRT.Routes).Count        
+
+        # list
+        $list = Get-AzRouteTable -ResourceGroupName $rgname
+        Assert-AreEqual 1 @($list).Count
+        Assert-AreEqual $list[0].ResourceGroupName $getRT.ResourceGroupName
+        Assert-AreEqual $list[0].Name $getRT.Name
+        Assert-AreEqual $list[0].DisablePeeringRoute $getRT.DisablePeeringRoute
+        Assert-AreEqual $list[0].Etag $getRT.Etag
+        Assert-AreEqual @($list[0].Routes).Count @($getRT.Routes).Count
+
+        # Update to "None"
+        $getRT.DisablePeeringRoute = "None"
+        $updatedRT = Set-AzRouteTable -RouteTable $getRT
+        Assert-AreEqual "None" $updatedRT.DisablePeeringRoute
+		
+        # Delete RouteTable
+        $delete = Remove-AzRouteTable -ResourceGroupName $rgname -name $routeTableName -PassThru -Force
+        Assert-AreEqual true $delete
+        
+        $list = Get-AzRouteTable -ResourceGroupName $rgname
+        Assert-AreEqual 0 @($list).Count
+    }
+    finally
+    {
+        # Cleanup
+        Clean-ResourceGroup $rgname
+    }
+}